Refactor: Restructure storage package
The current storage package uses TaskList, which is a model
representation of tasks during Ello's runtime. Hence, it's better to
move it to the model package.

Now, a TaskStorage class is used to handle all interactions between the
runtime and the data file.

SerializableTask, TasksSaverAndLoader and StorageException classes
has also been updated to reflect the changes in the storage package.
